Project completed
The completed project is a simple practice in HTML and CSS that involves building a single-page website for the fictional JLD Aveo+ Cafe. The webpage showcases a variety of coffee and desserts offered at the cafe, along with their corresponding prices.

Using HTML, the webpage is structured to provide a clear layout and organization. The different sections of the page, such as the header, menu items, and footer, are appropriately defined using HTML tags. This ensures proper semantics and enhances the accessibility and search engine optimization of the webpage.

CSS is applied to style the webpage and make it visually appealing. The use of CSS enables customization of the layout, typography, colors, and overall presentation of the content. Through carefully selected fonts, colors, and spacing, the design aims to create a pleasant and inviting atmosphere that aligns with the cafe's branding.

The project serves as a valuable exercise in combining HTML and CSS to create a responsive and well-designed single-page website. By showcasing coffee and dessert items along with their prices, the webpage effectively communicates the offerings of the JLD Aveo+ Cafe. It demonstrates the ability to apply HTML and CSS skills to develop a functional and visually appealing website, which are fundamental skills for web development.